+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.11] - Unreleased
-No changes yet.
+## Added
+- record per-frame FFRAC*TRANSP for ELG, BGS profiles.
 
 ## [0.1.10] - 2021-05-17
 ## Added